GCC 3 compilation fixes
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@9944 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/mail/MailComponent.cpp
@@ -518,7 +518,7 @@ BTextMailComponent::ParseRaw()
 	if (buffer == NULL)
 		return B_NO_MEMORY;
 
-	ssize_t bytes;
+	int32 bytes;
 	if ((bytes = raw_data->Read(buffer, raw_length)) < 0)
 		return B_IO_ERROR;
